- use u8 for entry point and pipeline index
- added documentation

-removed not implemented wgpu test

+++ candle-wgpu-kernels/build.rs
@@ -146,7 +146,7 @@ fn main() {
                     {}
                 }}
             }} 
-            pub fn from_index(index : u16) -> Self{{
+            pub fn from_index(index : u8) -> Self{{
                 match index{{
                     {}
                     _=> {{todo!()}}
